requirements.txt
python-dotenv>=1.0
chardet>=5.0
openpyxl>=3.1
openai>=1.0